Show/hide transition buttons according to permission
Instead of displaying all buttons and check if the user is allowed to
make the transition, first checks whether the user has permission and
then displays or hides the button. Transitions without permission
configured are not affected.
need to use the template processor
django.core.context_processors.request, put in
TEMPLATE_CONTEXT_PROCESSORS